+++ src/Api/Dfe.Complete.Api/Controllers/ProjectsController.cs
@@ -12,6 +12,9 @@ using Microsoft.AspNetCore.Authorization;
 using Dfe.Complete.Application.Projects.Commands.RemoveProject;
 using Dfe.Complete.Application.Projects.Commands.UpdateProject;
 using Dfe.Complete.Application.Projects.Queries.SearchProjects;
+using Dfe.Complete.Application.Notes.Queries;
+using Dfe.Complete.Application.Notes.Commands;
+using Dfe.Complete.Application.Common.Models;
 
 namespace Dfe.Complete.Api.Controllers
 {
@@ -146,7 +149,7 @@ namespace Dfe.Complete.Api.Controllers
             var project = await sender.Send(request, cancellationToken);
             return Ok(project.Value ?? []);
         }
-        
+
         /// <summary>
         /// Returns the number of Projects
         /// </summary>
@@ -349,6 +352,90 @@ namespace Dfe.Complete.Api.Controllers
             var project = await sender.Send(request, cancellationToken);
             return Ok(project.Value?.ProjectModels ?? []);
         }
+
+        /// <summary>
+        /// Returns a list of Notes for a Project
+        /// </summary>
+        /// <param name="request">The request.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        [Authorize(Policy = "CanRead")]
+        [HttpGet]
+        [Route("Notes")]
+        [SwaggerResponse(200, "Notes for project", typeof(List<NoteDto>))]
+        [SwaggerResponse(404, "Project not found")]
+        public async Task<IActionResult> GetNotesByProjectIdAsync([FromQuery] GetNotesByProjectIdQuery request, CancellationToken cancellationToken)
+        {
+            var result = await sender.Send(request, cancellationToken);
+
+            if (!result.IsSuccess || result.Value is null)
+                return NotFound();
+
+            return Ok(result.Value);
+        }
+
+        /// <summary>
+        /// Create a new Note for a Project
+        /// </summary>
+        /// <param name="request">The request.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        [Authorize(Policy = "CanReadWrite")]
+        [HttpPost]
+        [Route("Notes")]
+        [SwaggerResponse(201, "Note ID", typeof(NoteId))]
+        [SwaggerResponse(404, "Note not found")]
+        public async Task<IActionResult> CreateProjectNoteAsync([FromBody] CreateNoteCommand request, CancellationToken cancellationToken)
+        {
+            var result = await sender.Send(request, cancellationToken);
+            return Created("", result.Value);
+        }
+
+        /// <summary>
+        /// Update a Note for a Project
+        /// </summary>
+        /// <param name="request">The request.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        [Authorize(Policy = "CanReadWriteUpdate")]
+        [HttpPut]
+        [Route("Notes")]
+        [SwaggerResponse(200, "Note ID", typeof(NoteId))]
+        [SwaggerResponse(401, "Unauthorized access")]
+        [SwaggerResponse(404, "Note not found")]
+        public async Task<IActionResult> UpdateProjectNoteAsync([FromBody] UpdateNoteCommand request, CancellationToken cancellationToken)
+        {
+            var result = await sender.Send(request, cancellationToken);
+
+            if (!result.IsSuccess)
+            {
+                return result.ErrorType == ErrorType.NotFound
+                    ? NotFound(result.Error)
+                    : StatusCode(500);
+            }
+            return Ok(result.Value);
+        }
+
+        /// <summary>
+        /// Delete a Note for a Project
+        /// </summary>
+        /// <param name="request"></param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        [Authorize(Policy = "CanReadWriteUpdateDelete")]
+        [HttpDelete]
+        [Route("Notes")]
+        [SwaggerResponse(200, "Success", typeof(bool))]
+        [SwaggerResponse(404, "Project not found")]
+        [SwaggerResponse(404, "Note not found")]
+        public async Task<IActionResult> DeleteProjectNoteAsync([FromBody] RemoveNoteCommand request, CancellationToken cancellationToken)
+        {
+            var result = await sender.Send(request, cancellationToken);
+            if (!result.IsSuccess)
+            {
+                return result.ErrorType == ErrorType.NotFound
+                    ? NotFound(result.Error)
+                    : StatusCode(500);
+            }
+            return Ok(result.Value);
+        }
+
         /// <summary>
         /// Returns a list of all projects statistics.
         /// </summary>