+++ CHANGELOG.md
@@ -22,13 +22,17 @@ Security in case of vulnerabilities.
 
 ## Unreleased  
 
-### Added
+### Added  
+- Add notes repository, along with create, read, update and delete queries/commands
+- Add project notes page (`/projects/{projectId}/notes`)
+- Add project notes editing page (`/projects/{projectId}/notes/{noteId}/edit`)
+- Add project notes creation page (`/projects/{projectId}/notes/new`)
+- Add ability to delete note (`/projects/{projectId}/notes/{noteId}/delete`)
+- Attach user ID from DB as custom claim
 
 ### Changed  
 
 ### Fixed
-
-### Security
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-07-10.519...main) for everything awaiting release
 ---